+++ spark/src/main/scala-shims/spark-4.2/SetDSv2SchemaEvolutionShims.scala
@@ -16,15 +16,48 @@
 
 package org.apache.spark.sql.delta
 
+import org.apache.spark.sql.delta.sources.DeltaSQLConf
+
 import org.apache.spark.sql.SparkSession
+import org.apache.spark.sql.catalyst.plans.logical.{AppendData, InsertIntoStatement, LogicalPlan, MergeIntoTable, OverwriteByExpression, OverwritePartitionsDynamic}
+import org.apache.spark.sql.catalyst.rules.Rule
 
 /**
- * Spark 4.2 and above: pass-through shim that delegates to the actual [[SetDSv2SchemaEvolution]]
- * rule.
- *
- * This pre-resolution rule sets `withSchemaEvolution = true` on plan nodes when schema
+ * Pre-resolution rule that sets `withSchemaEvolution = true` on plan nodes when schema
  * evolution is enabled in Delta via either:
  * - the dataframe writer option `.option("mergeSchema", "true")`
  * - the session-level spark conf `spark.databricks.delta.schema.autoMerge.enabled`
+ *
+ * Spark only cares about `withSchemaEvolution`. This allows Spark to handle schema evolution for
+ * these write operations when using DSv2.
  */
-class SetDSv2SchemaEvolutionShims(session: SparkSession) extends SetDSv2SchemaEvolution(session)
+class SetDSv2SchemaEvolutionShims(session: SparkSession) extends Rule[LogicalPlan] {
+
+  private def canMergeSchema(writeOptions: Map[String, String]): Boolean =
+    new DeltaOptions(writeOptions, session.sessionState.conf).canMergeSchema
+
+  override def apply(plan: LogicalPlan): LogicalPlan = {
+    if (session.sessionState.conf.getConf(DeltaSQLConf.V2_ENABLE_MODE) == "NONE") return plan
+
+    // This rule runs during pre-resolution so it cannot identify the target table type. It applies
+    // to all DSv2 tables that declare AUTOMATIC_SCHEMA_EVOLUTION when a Delta-specific config is
+    // used to enable schema evolution.
+    // Note: the rule can't run later as it must run before [[ResolveSchemaEvolution]].
+    plan.resolveOperatorsDown {
+      case a: AppendData if !a.withSchemaEvolution && canMergeSchema(a.writeOptions) =>
+        a.copy(withSchemaEvolution = true)
+      case o: OverwriteByExpression
+          if !o.withSchemaEvolution && canMergeSchema(o.writeOptions) =>
+        o.copy(withSchemaEvolution = true)
+      case o: OverwritePartitionsDynamic
+          if !o.withSchemaEvolution && canMergeSchema(o.writeOptions) =>
+        o.copy(withSchemaEvolution = true)
+      // INSERT INTO (SQL) and MERGE don't support passing writer options, so only check the spark
+      // conf for schema evolution.
+      case i: InsertIntoStatement if !i.withSchemaEvolution && canMergeSchema(Map.empty) =>
+        i.copy(withSchemaEvolution = true)
+      case m: MergeIntoTable if !m.withSchemaEvolution && canMergeSchema(Map.empty) =>
+        m.copy(withSchemaEvolution = true)
+    }
+  }
+}
